removed unnecessary parameters to events, moved globals settings to a service, moved control structure for demographics auth error
