+++ src/Events/Appointments/CalendarFilterEvent.php
@@ -8,7 +8,6 @@
 
 namespace OpenEMR\Events\Appointments;
 
-use OpenEMR\Services\UserService;
 use Symfony\Component\EventDispatcher\Event;
 
 /**
@@ -28,12 +27,6 @@ class CalendarFilterEvent extends Event
     const EVENT_HANDLE = 'calendar.customFilter';
 
     /**
-     * @var null|UserService
-     *
-     */
-    private $userService = null;
-
-    /**
      * @var string
      *
      * This is the custom filter that can add to the calendar event query
@@ -41,23 +34,20 @@ class CalendarFilterEvent extends Event
     private $customWhereFilter = '1';
 
     /**
+     * Get an string representing the calendar event filter
      *
+     * @return string
      */
-    public function __construct(UserService $userService)
-    {
-        $this->userService = $userService;
-    }
-
-    public function getUserService()
-    {
-        return $this->userService;
-    }
-
     public function getCustomWhereFilter()
     {
         return $this->customWhereFilter;
     }
 
+    /**
+     * @param $customWhereFilter
+     *
+     * Add a custom filter to the WHERE clause of calendar event query
+     */
     public function setCustomWhereFilter($customWhereFilter)
     {
         $this->customWhereFilter = $customWhereFilter;
