+++ src/Events/Globals/GlobalsInitializedEvent.php
@@ -8,7 +8,7 @@
 
 namespace OpenEMR\Events\Globals;
 
-use OpenEMR\Services\UserService;
+use OpenEMR\Services\Globals\GlobalsService;
 use Symfony\Component\EventDispatcher\Event;
 
 /**
@@ -27,87 +27,28 @@ class GlobalsInitializedEvent extends Event
     const EVENT_HANDLE = 'globals.initialized';
 
     /**
-     * @var null|UserService
-     *
-     */
-    private $userService = null;
-
-    /**
      * @var array
      *
-     * The entire globals structure
+     * The globals service for manipulating globals data structure
      */
-    private $globalsMetadata = [];
+    private $globalsService = null;
 
     /**
-     * @var array
+     * GlobalsInitializedEvent constructor.
+     * @param array $globalsService
+     */
+    public function __construct(GlobalsService $globalsService)
+    {
+        $this->globalsService = $globalsService;
+    }
+
+    /**
+     * @return GlobalsService
      *
-     * User-specific globals
+     * Get the instance of globals service
      */
-    private $userSpecificGlobals = [];
-
-    /**
-     * @var array
-     *
-     *  User specific tabs
-     */
-    private $userSpecificTabs = [];
-
-    /**
-     * GlobalsInitEvent constructor.
-     * @param $userService
-     * @param $GLOBALS_METADATA
-     * @param $USER_SPECIFIC_GLOBALS
-     * @param $USER_SPECIFIC_TABS
-     */
-    public function __construct($userService, $GLOBALS_METADATA, $USER_SPECIFIC_GLOBALS, $USER_SPECIFIC_TABS)
+    public function getGlobalsService()
     {
-        $this->userService = $userService;
-        $this->globalsMetadata = $GLOBALS_METADATA;
-        $this->userSpecificGlobals = $USER_SPECIFIC_GLOBALS;
-        $this->userSpecificTabs = $USER_SPECIFIC_TABS;
-    }
-
-    public function getUserService()
-    {
-        return $this->userService;
-    }
-
-    public function save()
-    {
-        global $GLOBALS_METADATA, $USER_SPECIFIC_GLOBALS, $USER_SPECIFIC_TABS;
-        $GLOBALS_METADATA = $this->globalsMetadata;
-        $USER_SPECIFIC_GLOBALS = $this->userSpecificGlobals;
-        $USER_SPECIFIC_TABS = $this->userSpecificTabs;
-    }
-
-    public function createSection($section, $beforeSection = false)
-    {
-        if (!isset($this->globalsMetadata[$section])) {
-            if ($beforeSection !== false &&
-                isset($this->globalsMetadata[$beforeSection])) {
-                $beforeSectionIndex = array_search($beforeSection, array_keys($this->globalsMetadata));
-                $this->globalsMetadata = array_slice($this->globalsMetadata, 0, $beforeSectionIndex, true) +
-                    array($section => []) +
-                    array_slice($this->globalsMetadata, $beforeSectionIndex, count($this->globalsMetadata) - 1, true) ;
-            } else {
-                $this->globalsMetadata[$section] = [];
-            }
-        } else {
-            throw new \Exception("Section already exists and cannot be created");
-        }
-    }
-
-    public function appendToSection($section, $key, GlobalSetting $global)
-    {
-        $this->globalsMetadata[$section][$key] = $global->format();
-        if ($global->isUserSetting()) {
-            $this->userSpecificGlobals[]= $key;
-        }
-    }
-
-    public function getGlobalsMetadata()
-    {
-        return $this->globalsMetadata;
+        return $this->globalsService;
     }
 }
